Fix duplicate alert
Right now two alerts appear when the browser doesn't support push notifications. Should be one.
